Move Hub classes to its own subpackage
