+++ src/meltano/core/hub/__init__.py
@@ -1,0 +1,5 @@
+"""Clients and APIs for interacting with Meltano Hub."""
+
+from .client import MeltanoHubService
+
+__all__ = ["MeltanoHubService"]
